+++ test/config-sample.js
@@ -1,13 +1,7 @@
+'use strict';
+
 module.exports = {
-	mongo: {
-		//		database: 'diaspora_test',
-		//		username: 'user',
-		//		password: false,
-	},
-	redis: {
-		//	database: 1,
-	},
-	localstorage: {
-		//		data_dir: '.localStorageTest'
-	},
+	mongo:        {},
+	redis:        {},
+	localstorage: {},
 };